Finished 2016/22
Since the actual grid described by my input data was pretty
straightforward, the solution was pretty easy to write.
All the nodes are either unmovable obstacles, movable data,
or empty (only one is empty). No node has enough space
for two nodes' data. Basically we just have to move the single
space around, avoiding the obstacles.
The obstacles are all in a horizontal line, in one row,
nowhere near the top.
All way have to do is take the empty space past the obstacles,
up to the prize data, and use it to step the prize leftwards
until we get to the destination.