2018/19 parts 1 and 2
I find this kind of aoc question a bit unsatisfying, because
the solution is to look at my own input until I figure out
what algorithm it is performing, and then skip it and calculate
the answer a quicker way. So if my input was significantly
different, then this program wouldn't work.
But that seems to be the only way to solve it, unless I knew
what everyone else's input looked like as well.