2019/15
There's probably a simpler way.
Explore the maze, trying direction in every room until everything's been explored.
Use a steps grid to find routes between rooms.
When the grid is all filled in, the value of the steps grid at the target
is the number of steps to the target.
Then for part 2, recalculate the steps grid, with the target (the oxygen
generator) at step zero. The maximum value now in the steps grid
is the number of minutes the maze takes to fill with oxygen.